# cogniware-opea-ims/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cogniware Core API
  description: |
    High-Performance LLM Acceleration Platform
    
    ## Performance
    - **15.4x** average speed improvement
    - **8.5ms** inference latency
    - **60,000** tokens/second (4 LLMs in parallel)
    
    ## Features
    - Custom GPU drivers with direct hardware access
    - Multi-LLM orchestration (4+ models simultaneously)
    - Complete MCP integration (10 subsystems, 50+ tools)
    - Enterprise security (JWT, OAuth2, MFA, RBAC)
    
  version: 1.0.0
  contact:
    name: Cogniware Support
    email: support@cogniware.ai
    url: https://cogniware.ai
  license:
    name: Proprietary
    url: https://cogniware.ai/license
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.cogniware.ai/v1
    description: Production server
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate with username and password to receive access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                      expires_in:
                        type: integer
                        example: 3600
                      refresh_token:
                        type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []